Remove ConfigurationReader/Configuration distinction
The reader only contained a couple of static methods. They can simply be folded
into Configuration itself.
