+++ server/src/main/java/com/clarkparsia/pellet/server/Configuration.java
@@ -1,5 +1,7 @@
 package com.clarkparsia.pellet.server;
 
+import java.util.Properties;
+
 /**
  * @author Edgar Rodriguez-Diaz
  */
@@ -16,4 +18,26 @@ public class Configuration {
 	public static String PELLET_PORT = "pellet.port";
 	public static String PELLET_UPDATE_INTERVAL = "pellet.update.interval.sec";
 	public static String PELLET_MANAGEMENT_PASSWORD = "pellet.password";
+
+	public static String getProperty(Properties properties, String key, String defaultValue) {
+		String val = properties.getProperty(key, defaultValue);
+		if (val == null) {
+			throw new IllegalArgumentException("Value of configuration property " + key + " is missing");
+		}
+		return val;
+	}
+
+	public static int getPropertyAsInteger(Properties properties, String key, int defaultValue) {
+		String val = properties.getProperty(key);
+		if (val == null) {
+			return defaultValue;
+		}
+
+		try {
+			return Integer.parseInt(val);
+		}
+		catch (NumberFormatException e) {
+			throw new IllegalArgumentException("Value of configuration property " + key + " is not a valid integer: " + defaultValue);
+		}
+	}
 }
