+++ server/src/main/java/com/clarkparsia/pellet/server/PelletSettings.java
@@ -17,11 +17,11 @@ public class PelletSettings {
     }
 
     public String home() {
-        return ConfigurationReader.getProperty(settings, Configuration.PELLET_HOME, System.getProperty("user.dir"));
+        return Configuration.getProperty(settings, Configuration.PELLET_HOME, System.getProperty("user.dir"));
     }
 
     public String host() {
-        return ConfigurationReader.getProperty(settings, Configuration.PELLET_HOST, PelletServer.DEFAULT_HOST);
+        return Configuration.getProperty(settings, Configuration.PELLET_HOST, PelletServer.DEFAULT_HOST);
     }
 
     public String endpoint() {
@@ -29,15 +29,15 @@ public class PelletSettings {
     }
 
     public int port() {
-        return ConfigurationReader.getPropertyAsInteger(settings, Configuration.PELLET_PORT, PelletServer.DEFAULT_PORT);
+        return Configuration.getPropertyAsInteger(settings, Configuration.PELLET_PORT, PelletServer.DEFAULT_PORT);
     }
 
     public int updateIntervalInSeconds() {
-        return ConfigurationReader.getPropertyAsInteger(settings, Configuration.PELLET_UPDATE_INTERVAL, UPDATE_INTERVAL_DEFAULT_IN_SECONDS);
+        return Configuration.getPropertyAsInteger(settings, Configuration.PELLET_UPDATE_INTERVAL, UPDATE_INTERVAL_DEFAULT_IN_SECONDS);
     }
 
 	public String managementPassword() {
-		final String pass = ConfigurationReader.getProperty(settings, Configuration.PELLET_MANAGEMENT_PASSWORD, null);
+		final String pass = Configuration.getProperty(settings, Configuration.PELLET_MANAGEMENT_PASSWORD, null);
 		if (pass == null) {
 			throw new RuntimeException("Unavilable management password");
 		}
