+++ server/src/main/java/com/clarkparsia/pellet/server/ProtegeSettings.java
@@ -23,23 +23,23 @@ public class ProtegeSettings {
     }
 
     public String host() {
-        return ConfigurationReader.getProperty(settings, Configuration.PROTEGE_HOST, DEFAULT_PROTEGE_HOST);
+        return Configuration.getProperty(settings, Configuration.PROTEGE_HOST, DEFAULT_PROTEGE_HOST);
     }
 
     public int port() {
-        return ConfigurationReader.getPropertyAsInteger(settings, Configuration.PROTEGE_PORT, DEFAULT_PROTEGE_PORT);
+        return Configuration.getPropertyAsInteger(settings, Configuration.PROTEGE_PORT, DEFAULT_PROTEGE_PORT);
     }
 
     public String username() {
-        return ConfigurationReader.getProperty(settings, Configuration.PROTEGE_USERNAME, null);
+        return Configuration.getProperty(settings, Configuration.PROTEGE_USERNAME, null);
     }
 
     public String password() {
-        return ConfigurationReader.getProperty(settings, Configuration.PROTEGE_PASSWORD, null);
+        return Configuration.getProperty(settings, Configuration.PROTEGE_PASSWORD, null);
     }
 
     public Set<String> ontologies() {
-        final String ontologiesList = ConfigurationReader.getProperty(settings, Configuration.PROTEGE_ONTOLOGIES, "");
+        final String ontologiesList = Configuration.getProperty(settings, Configuration.PROTEGE_ONTOLOGIES, "");
 
         // try parsing the ontology names list
         return ImmutableSet.copyOf(Splitter.on(',').omitEmptyStrings().trimResults().split(ontologiesList));
